+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # multi-image-mergetool changelog
+1.12.0 - Added server tests for /update-image-set/:filepath
+
+1.11.1 - Added server tests for /images/:filepath
+
+1.11.0 - Added initial server tests
+
 1.10.0 - Added tests for "Find similar images" and repaired upsizing of small images in similar results
 
 1.9.1 - Added screenshot uploading for Travis CI and repaired CI tests via compiling first